update mosquitto lib to 1.1.2

+++ libmosquitto/send_mosq.c
@@ -1,31 +1,31 @@
 /*
-Copyright (c) 2009-2011 Roger Light <roger@atchoo.org>
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-
-1. Redistributions of source code must retain the above copyright notice,
-   this list of conditions and the following disclaimer.
-2. Redistributions in binary form must reproduce the above copyright
-   notice, this list of conditions and the following disclaimer in the
-   documentation and/or other materials provided with the distribution.
-3. Neither the name of mosquitto nor the names of its
-   contributors may be used to endorse or promote products derived from
-   this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
-LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-POSSIBILITY OF SUCH DAMAGE.
-*/
+ Copyright (c) 2009-2011 Roger Light <roger@atchoo.org>
+ All rights reserved.
+ 
+ Redistribution and use in source and binary forms, with or without
+ modification, are permitted provided that the following conditions are met:
+ 
+ 1. Redistributions of source code must retain the above copyright notice,
+ this list of conditions and the following disclaimer.
+ 2. Redistributions in binary form must reproduce the above copyright
+ notice, this list of conditions and the following disclaimer in the
+ documentation and/or other materials provided with the distribution.
+ 3. Neither the name of mosquitto nor the names of its
+ contributors may be used to endorse or promote products derived from
+ this software without specific prior written permission.
+ 
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+ LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ POSSIBILITY OF SUCH DAMAGE.
+ */
 
 #include <assert.h>
 #include <stdio.h>
@@ -94,7 +94,7 @@ int _mosquitto_send_pubcomp(struct mosquitto *mosq, uint16_t mid)
 int _mosquitto_send_publish(struct mosquitto *mosq, uint16_t mid, const char *topic, uint32_t payloadlen, const void *payload, int qos, bool retain, bool dup)
 {
 #ifdef WITH_BROKER
-	int len;
+	size_t len;
 #ifdef WITH_BRIDGE
 	int i;
 	struct _mqtt3_bridge_topic *cur_topic;
@@ -106,7 +106,7 @@ int _mosquitto_send_publish(struct mosquitto *mosq, uint16_t mid, const char *to
 #endif
 	assert(mosq);
 	assert(topic);
-
+    
 	if(mosq->sock == INVALID_SOCKET) return MOSQ_ERR_NO_CONN;
 #ifdef WITH_BROKER
 	if(mosq->listener && mosq->listener->mount_point){
@@ -124,7 +124,7 @@ int _mosquitto_send_publish(struct mosquitto *mosq, uint16_t mid, const char *to
 			cur_topic = &mosq->bridge->topics[i];
 			if(cur_topic->remote_prefix || cur_topic->local_prefix){
 				/* Topic mapping required on this topic if the message matches */
-
+                
 				rc = mosquitto_topic_matches_sub(cur_topic->local_topic, topic, &match);
 				if(rc){
 					return rc;
@@ -143,7 +143,7 @@ int _mosquitto_send_publish(struct mosquitto *mosq, uint16_t mid, const char *to
 							mapped_topic = topic_temp;
 						}
 					}
-
+                    
 					if(cur_topic->remote_prefix){
 						/* This prefix needs adding. */
 						len = strlen(mapped_topic) + strlen(cur_topic->remote_prefix)+1;
@@ -171,7 +171,7 @@ int _mosquitto_send_publish(struct mosquitto *mosq, uint16_t mid, const char *to
 #else
 	_mosquitto_log_printf(mosq, MOSQ_LOG_DEBUG, "Client %s sending PUBLISH (d%d, q%d, r%d, m%d, '%s', ... (%ld bytes))", mosq->id, dup, qos, retain, mid, topic, (long)payloadlen);
 #endif
-
+    
 	return _mosquitto_send_real_publish(mosq, mid, topic, payloadlen, payload, qos, retain, dup);
 }
 
@@ -200,11 +200,11 @@ int _mosquitto_send_command_with_mid(struct mosquitto *mosq, uint8_t command, ui
 {
 	struct _mosquitto_packet *packet = NULL;
 	int rc;
-
+    
 	assert(mosq);
 	packet = _mosquitto_calloc(1, sizeof(struct _mosquitto_packet));
 	if(!packet) return MOSQ_ERR_NOMEM;
-
+    
 	packet->command = command;
 	if(dup){
 		packet->command |= 8;
@@ -215,10 +215,10 @@ int _mosquitto_send_command_with_mid(struct mosquitto *mosq, uint8_t command, ui
 		_mosquitto_free(packet);
 		return rc;
 	}
-
+    
 	packet->payload[packet->pos+0] = MOSQ_MSB(mid);
 	packet->payload[packet->pos+1] = MOSQ_LSB(mid);
-
+    
 	return _mosquitto_packet_queue(mosq, packet);
 }
 
@@ -227,20 +227,20 @@ int _mosquitto_send_simple_command(struct mosquitto *mosq, uint8_t command)
 {
 	struct _mosquitto_packet *packet = NULL;
 	int rc;
-
+    
 	assert(mosq);
 	packet = _mosquitto_calloc(1, sizeof(struct _mosquitto_packet));
 	if(!packet) return MOSQ_ERR_NOMEM;
-
+    
 	packet->command = command;
 	packet->remaining_length = 0;
-
+    
 	rc = _mosquitto_packet_alloc(packet);
 	if(rc){
 		_mosquitto_free(packet);
 		return rc;
 	}
-
+    
 	return _mosquitto_packet_queue(mosq, packet);
 }
 
@@ -249,15 +249,15 @@ int _mosquitto_send_real_publish(struct mosquitto *mosq, uint16_t mid, const cha
 	struct _mosquitto_packet *packet = NULL;
 	int packetlen;
 	int rc;
-
+    
 	assert(mosq);
 	assert(topic);
-
+    
 	packetlen = 2+strlen(topic) + payloadlen;
 	if(qos > 0) packetlen += 2; /* For message id */
 	packet = _mosquitto_calloc(1, sizeof(struct _mosquitto_packet));
 	if(!packet) return MOSQ_ERR_NOMEM;
-
+    
 	packet->mid = mid;
 	packet->command = PUBLISH | ((dup&0x1)<<3) | (qos<<1) | retain;
 	packet->remaining_length = packetlen;
@@ -271,11 +271,11 @@ int _mosquitto_send_real_publish(struct mosquitto *mosq, uint16_t mid, const cha
 	if(qos > 0){
 		_mosquitto_write_uint16(packet, mid);
 	}
-
+    
 	/* Payload */
 	if(payloadlen){
 		_mosquitto_write_bytes(packet, payload, payloadlen);
 	}
-
+    
 	return _mosquitto_packet_queue(mosq, packet);
 }
